Removed Monk and Mongolia for MongoDB driver wrapper due to limitation of bulkwrite.
MongoDB native node driver provides a function bulkwrite for ordered or non-ordered query (in my case is insertion), but Monk and Mongolia don't support bulkwrite. One problem of using native driver is to pass along an open-connection. MongoClient.connect or connection.open are async function, which takes callback to return a connection object, the server script need to be run after this connection is established to pass along the connection for routing.